requirement-win.txt
apipkg==1.5
Appium-Python-Client==0.44
atomicwrites==1.2.1
attrs==18.2.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
comtypes==1.1.7
configparser==3.5.0
cycler==0.10.0
execnet==1.7.1
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
more-itertools==4.3.0
numpy==1.16.4
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
pluggy==0.9.0
py==1.7.0
PyMySQL==0.9.3
pyparsing==2.4.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
pytest==4.4.1
pytest-assume==1.2.2
pytest-forked==1.0.2
pytest-html==1.20.0
pytest-metadata==1.8.0
pytest-ordering==0.6
pytest-rerunfailures==7.0
pytest-xdist==1.29.0
python-dateutil==2.8.0
pywin32==224
pywinauto==0.6.7
PyYAML==5.3.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.3
